Renames the blob loops (Entities -> Blobs)
