+++ src/main/java/sirius/biz/storage/layer2/mongo/RemoveDeletedBlobsLoop.java
@@ -6,10 +6,11 @@
  * http://www.scireum.de - info@scireum.de
  */
 
-package sirius.biz.storage.layer2.jdbc;
+package sirius.biz.storage.layer2.mongo;
 
 import sirius.biz.storage.util.StorageUtils;
-import sirius.db.jdbc.OMA;
+import sirius.db.mongo.Mango;
+import sirius.db.mongo.Mongo;
 import sirius.kernel.async.BackgroundLoop;
 import sirius.kernel.commons.Strings;
 import sirius.kernel.di.std.Part;
@@ -18,24 +19,26 @@ import sirius.kernel.health.Exceptions;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.sql.SQLException;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * Finally deletes {@link SQLDirectory directories} and {@link SQLBlob blobs} which have been marked as deleted.
+ * Finally deletes {@link MongoDirectory directories} and {@link MongoBlob blobs} which have been marked as deleted.
  */
-@Register(framework = SQLBlobStorage.FRAMEWORK_JDBC_BLOB_STORAGE)
-public class RemoveDeletedEntitiesLoop extends BackgroundLoop {
+@Register(framework = MongoBlobStorage.FRAMEWORK_MONGO_BLOB_STORAGE)
+public class RemoveDeletedBlobsLoop extends BackgroundLoop {
 
     private static final double FREQUENCY_EVERY_FIFTEEN_SECONDS = 1 / 15d;
 
     @Part
-    private OMA oma;
+    private Mango mango;
+
+    @Part
+    private Mongo mongo;
 
     @Nonnull
     @Override
     public String getName() {
-        return "storage-layer2-delete";
+        return "storage-layer2-blob-delete";
     }
 
     @Override
@@ -58,20 +61,20 @@ public class RemoveDeletedEntitiesLoop extends BackgroundLoop {
 
     private AtomicInteger deleteBlobs() {
         AtomicInteger numBlobs = new AtomicInteger();
-        oma.select(SQLBlob.class).eq(SQLBlob.DELETED, true).limit(256).iterateAll(blob -> {
+        mango.select(MongoBlob.class).eq(MongoBlob.DELETED, true).limit(256).iterateAll(blob -> {
             try {
                 if (Strings.isFilled(blob.getPhysicalObjectKey())) {
                     blob.getStorageSpace().getPhysicalSpace().delete(blob.getPhysicalObjectKey());
                 }
 
-                oma.delete(blob);
+                mango.delete(blob);
                 numBlobs.incrementAndGet();
             } catch (Exception e) {
                 Exceptions.handle()
                           .to(StorageUtils.LOG)
                           .error(e)
                           .withSystemErrorMessage(
-                                  "Layer 2/SQL: Failed to finally delete the blob %s (%s) in %s: %s (%s)",
+                                  "Layer 2/Mongo: Failed to finally delete the blob %s (%s) in %s: %s (%s)",
                                   blob.getBlobKey(),
                                   blob.getFilename(),
                                   blob.getSpaceName())
@@ -84,16 +87,16 @@ public class RemoveDeletedEntitiesLoop extends BackgroundLoop {
 
     private AtomicInteger deleteDirectories() {
         AtomicInteger numDirectories = new AtomicInteger();
-        oma.select(SQLDirectory.class).eq(SQLDirectory.DELETED, true).limit(256).iterateAll(dir -> {
+        mango.select(MongoDirectory.class).eq(MongoDirectory.DELETED, true).limit(256).iterateAll(dir -> {
             try {
                 propagateDelete(dir);
-                oma.delete(dir);
+                mango.delete(dir);
             } catch (Exception e) {
                 Exceptions.handle()
                           .to(StorageUtils.LOG)
                           .error(e)
                           .withSystemErrorMessage(
-                                  "Layer 2/SQL: Failed to finally delete the directory %s (%s) in %s: %s (%s)",
+                                  "Layer 2/Mongo: Failed to finally delete the directory %s (%s) in %s: %s (%s)",
                                   dir.getId(),
                                   dir.getName(),
                                   dir.getSpaceName())
@@ -105,14 +108,11 @@ public class RemoveDeletedEntitiesLoop extends BackgroundLoop {
         return numDirectories;
     }
 
-    private void propagateDelete(SQLDirectory dir) throws SQLException {
-        oma.updateStatement(SQLDirectory.class)
-           .set(SQLDirectory.DELETED, true)
-           .where(SQLDirectory.PARENT, dir.getId())
-           .executeUpdate();
-        oma.updateStatement(SQLBlob.class)
-           .set(SQLBlob.DELETED, true)
-           .where(SQLBlob.PARENT, dir.getId())
-           .executeUpdate();
+    private void propagateDelete(MongoDirectory dir) {
+        mongo.update()
+             .set(MongoDirectory.DELETED, true)
+             .where(MongoDirectory.PARENT, dir.getId())
+             .executeFor(MongoDirectory.class);
+        mongo.update().set(MongoBlob.DELETED, true).where(MongoBlob.PARENT, dir.getId()).executeFor(MongoBlob.class);
     }
 }
